// src/utils/spotify-types.ts
export type TrackItem = Track | Episode;
// ...
export interface SimplifiedAlbum extends AlbumBase {
  album_group: string;
  artists: SimplifiedArtist[];
}

export interface SavedAlbum {
  added_at: string;
  album: Album;
}

export interface Album extends AlbumBase {
  artists: Artist[];
  tracks: Page<SimplifiedTrack>;
}

export interface Albums {
  albums: Album[];
}

export interface NewReleases {
  albums: Page<SimplifiedAlbum>;
}

export interface Copyright {
  text: string;
  type: string;
}

export interface Page<TItemType> {
  href: string;
  items: TItemType[];
  limit: number;
  next: string | null;
  offset: number;
  previous: string | null;
  total: number;
}

export interface PlaylistedTrack<Item extends TrackItem = TrackItem> {
  added_at: string;
  added_by: AddedBy;
  is_local: boolean;
  primary_color: string;
  track: Item;
}

export interface AddedBy {
  external_urls: ExternalUrls;
  href: string;
  id: string;
  type: string;
  uri: string;
}

export interface LinkedFrom {
  external_urls: ExternalUrls;
  href: string;
  id: string;
  type: string;
  uri: string;
}

export interface SimplifiedTrack {
  artists: SimplifiedArtist[];
  available_markets: string[];
  disc_number: number;
  duration_ms: number;
  episode: boolean;
  explicit: boolean;
  external_urls: ExternalUrls;
  href: string;
  id: string;
  is_local: boolean;
  name: string;
  preview_url: string | null;
  track: boolean;
  track_number: number;
  type: string;
  uri: string;
  is_playable?: boolean;
  linked_from?: LinkedFrom;
  restrictions?: Restrictions;
}

export interface SavedTrack {
  added_at: string;
  track: Track;
}

export interface ExternalIds {
  isrc: string;
  ean: string;
  upc: string;
}

export interface Track extends SimplifiedTrack {
  album: SimplifiedAlbum;
  external_ids: ExternalIds;
  popularity: number;
}
// ...
export interface Followers {
  href: string | null;
  total: number;
}

export interface ExternalUrls {
  spotify: string;
}
// ...
export interface Image {
  url: string;
  height: number;
  width: number;
}
// ...
export interface SimplifiedEpisode {
  audio_preview_url: string;
  description: string;
  html_description: string;
  duration_ms: number;
  explicit: boolean;
  external_urls: ExternalUrls;
  href: string;
  id: string;
  images: Image[];
  is_externally_hosted: boolean;
  is_playable: boolean;
  language: string;
  languages: string[];
  name: string;
  release_date: string;
  release_date_precision: string;
  resume_point: ResumePoint;
  type: string;
  uri: string;
  restrictions: Restrictions;
}

export interface Episode extends SimplifiedEpisode {
  show: SimplifiedShow;
}

export interface SavedEpisode {
  added_at: string;
  episode: Episode;
}

export interface SimplifiedShow {
  available_markets: string[];
  copyrights: Copyright[];
  description: string;
  html_description: string;
  explicit: boolean;
  external_urls: ExternalUrls;
  href: string;
  id: string;
  images: Image[];
  is_externally_hosted: boolean;
  languages: string[];
  media_type: string;
  name: string;
  publisher: string;
  type: string;
  uri: string;
  total_episodes: number;
}

export interface SavedShow {
  added_at: string;
  show: SimplifiedShow;
}

export interface Show extends SimplifiedShow {
  episodes: Page<SimplifiedEpisode>;
}

export interface SnapshotReference {
  snapshot_id: string;
}

interface PlaylistBase {
  collaborative: boolean;
  description: string;
  external_urls: ExternalUrls;
  followers: Followers;
  href: string;
  id: string;
  images: Image[];
  name: string;
  owner: UserReference;
  primary_color: string;
  public: boolean;
  snapshot_id: string;
  type: string;
  uri: string;
}

export interface Playlist<Item extends TrackItem = TrackItem> extends PlaylistBase {
  tracks: Page<PlaylistedTrack<Item>>;
}

export interface FeaturedPlaylists {
  message: string;
  playlists: Page<SimplifiedPlaylist>;
}

export interface SimplifiedPlaylist extends PlaylistBase {
  tracks: TrackReference | null;
}

export interface TrackReference {
  href: string;
  total: number;
}

export interface UserReference {
  display_name: string;
  external_urls: ExternalUrls;
  href: string;
  id: string;
  type: string;
  uri: string;
}